Updated Octopus Deploy resources to be v3-compliant.
The main change is that Path is no longer passed to the ConfigFile resource; this is instead determined at execution time by a combination of the tentacle name, home directory, and what's in the registry on a machine where the tentacle name already exists.  HomeDirectory is now a required property, and is a composite key along with TentacleName.

Octopus v3, by default, puts config files in a different location than previous versions.  The resource has been updated to mimic this behavior, except in situations where a tentacle / config file already exists in the old location (to avoid regenerating certificates / etc and breaking communication with the server.)